Issue 34: Move PhotoBundle into WebBundle
