--- src/Khatovar/Bundle/PhotoBundle/Resources/config/form_types.yml
@@ -1,9 +1,0 @@
-services:
-    khatovar_photo.form.type.photo:
-        class: Khatovar\Bundle\PhotoBundle\Form\Type\PhotoType
-        arguments:
-            - '@security.authorization_checker'
-            - '@khatovar_photo.form.subscriber.add_fields'
-            - '@khatovar_photo.form.subscriber.add_authorized'
-        tags:
-            - { name: form.type }
