--- src/Khatovar/Bundle/PhotoBundle/Resources/config/handlers.yml
@@ -1,7 +1,0 @@
-services:
-    khatovar_photo.handler.photo:
-        class: Khatovar\Bundle\PhotoBundle\Handler\PhotoHandler
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '@khatovar_photo.manager.photo'
-            - '@router'
