--- src/Khatovar/Bundle/PhotoBundle/Resources/config/routing.yml
@@ -1,26 +1,0 @@
-khatovar_web_photo:
-    path:       /
-    defaults:   { _controller: KhatovarPhotoBundle:Photo:index }
-
-khatovar_web_photo_new:
-    path:       /new
-    defaults:   { _controller: KhatovarPhotoBundle:Photo:new }
-
-khatovar_web_photo_create:
-    path:     /create
-    defaults: { _controller: KhatovarPhotoBundle:Photo:create }
-    methods:  [POST]
-
-khatovar_web_photo_edit:
-    path:       /{id}/edit
-    defaults:   { _controller: KhatovarPhotoBundle:Photo:edit }
-
-khatovar_web_photo_update:
-    path:     /{id}/update
-    defaults: { _controller: KhatovarPhotoBundle:Photo:update }
-    methods:  [POST, PUT]
-
-khatovar_web_photo_delete:
-    path:       /{id}/delete
-    defaults:   { _controller: KhatovarPhotoBundle:Photo:delete }
-    methods:  [POST, DELETE]
