+++ src/Khatovar/Bundle/WebBundle/Entity/Exaction.php
@@ -22,7 +22,7 @@
 namespace Khatovar\Bundle\WebBundle\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
-use Khatovar\Bundle\PhotoBundle\Entity\Photo;
+use Khatovar\Bundle\WebBundle\Entity\Photo;
 
 /**
  * @author Damien Carcel <damien.carcel@gmail.com>
