+++ src/Khatovar/Bundle/WebBundle/Resources/config/doctrine/Appearance.orm.yml
@@ -41,5 +41,5 @@ Khatovar\Bundle\WebBundle\Entity\Appearance:
             nullable: false
     oneToMany:
         photos:
-            targetEntity: Khatovar\Bundle\PhotoBundle\Entity\Photo
+            targetEntity: Khatovar\Bundle\WebBundle\Entity\Photo
             mappedBy: appearance
