+++ src/Khatovar/Bundle/WebBundle/Resources/config/form_types.yml
@@ -44,3 +44,12 @@ services:
             - '@khatovar_member.form.subscriber.remove_owner'
         tags:
             - { name: form.type }
+
+    khatovar_photo.form.type.photo:
+        class: Khatovar\Bundle\WebBundle\Form\Type\PhotoType
+        arguments:
+            - '@security.authorization_checker'
+            - '@khatovar_photo.form.subscriber.add_fields'
+            - '@khatovar_photo.form.subscriber.add_authorized'
+        tags:
+            - { name: form.type }
