+++ src/Khatovar/Bundle/WebBundle/Resources/config/routing/photo.yml
@@ -1,0 +1,26 @@
+khatovar_web_photo:
+    path:       /
+    defaults:   { _controller: KhatovarWebBundle:Photo:index }
+
+khatovar_web_photo_new:
+    path:       /new
+    defaults:   { _controller: KhatovarWebBundle:Photo:new }
+
+khatovar_web_photo_create:
+    path:     /create
+    defaults: { _controller: KhatovarWebBundle:Photo:create }
+    methods:  [POST]
+
+khatovar_web_photo_edit:
+    path:       /{id}/edit
+    defaults:   { _controller: KhatovarWebBundle:Photo:edit }
+
+khatovar_web_photo_update:
+    path:     /{id}/update
+    defaults: { _controller: KhatovarWebBundle:Photo:update }
+    methods:  [POST, PUT]
+
+khatovar_web_photo_delete:
+    path:       /{id}/delete
+    defaults:   { _controller: KhatovarWebBundle:Photo:delete }
+    methods:  [POST, DELETE]
