+++ src/Khatovar/Bundle/WebBundle/Resources/views/Photo/edit.html.twig
@@ -13,7 +13,7 @@
         {{ form_start(edit_form) }}
         {{ form_errors(edit_form) }}
 
-        {% include 'KhatovarPhotoBundle:Photo:_form.html.twig' with {'form': edit_form} %}
+        {% include 'KhatovarWebBundle:Photo:_form.html.twig' with {'form': edit_form} %}
 
         {{ form_rest(edit_form) }}
         {{ form_end(edit_form) }}
